refactor: split app into components and hooks
